move compress utils to bloomshipper
