+++ pkg/storage/stores/shipper/bloomshipper/compress_utils.go
@@ -1,4 +1,4 @@
-package bloomcompactor
+package bloomshipper
 
 import (
 	"fmt"
@@ -11,11 +11,10 @@ import (
 	"github.com/go-kit/log/level"
 
 	v1 "github.com/grafana/loki/pkg/storage/bloom/v1"
-	"github.com/grafana/loki/pkg/storage/stores/shipper/bloomshipper"
 )
 
-func compressBloomBlock(ref bloomshipper.BlockRef, archivePath, localDst string, logger log.Logger) (bloomshipper.Block, error) {
-	blockToUpload := bloomshipper.Block{}
+func CompressBloomBlock(ref BlockRef, archivePath, localDst string, logger log.Logger) (Block, error) {
+	blockToUpload := Block{}
 	archiveFile, err := os.Create(archivePath)
 	if err != nil {
 		return blockToUpload, err
@@ -32,8 +31,7 @@ func compressBloomBlock(ref bloomshipper.BlockRef, archivePath, localDst string,
 	return blockToUpload, nil
 }
 
-// TODO common with block_downloader, extract under bloomutils
-func uncompressBloomBlock(block *bloomshipper.LazyBlock, workingDirectory string) (string, error) {
+func UncompressBloomBlock(block *LazyBlock, workingDirectory string, logger log.Logger) (string, error) {
 	workingDirectoryPath := filepath.Join(workingDirectory, block.BlockPath)
 	err := os.MkdirAll(workingDirectoryPath, os.ModePerm)
 	if err != nil {
@@ -45,7 +43,9 @@ func uncompressBloomBlock(block *bloomshipper.LazyBlock, workingDirectory string
 	}
 	defer func() {
 		os.Remove(archivePath)
-		// todo log err
+		if err != nil {
+			level.Error(logger).Log("msg", "removing archive file", "err", err, "file", archivePath)
+		}
 	}()
 	err = extractArchive(archivePath, workingDirectoryPath)
 	if err != nil {
@@ -54,8 +54,7 @@ func uncompressBloomBlock(block *bloomshipper.LazyBlock, workingDirectory string
 	return workingDirectoryPath, nil
 }
 
-// TODO common with block_downloader, extract under bloomutils
-func writeDataToTempFile(workingDirectoryPath string, block *bloomshipper.LazyBlock) (string, error) {
+func writeDataToTempFile(workingDirectoryPath string, block *LazyBlock) (string, error) {
 	defer block.Data.Close()
 	archivePath := filepath.Join(workingDirectoryPath, block.BlockPath[strings.LastIndex(block.BlockPath, "/")+1:])
 
@@ -71,7 +70,6 @@ func writeDataToTempFile(workingDirectoryPath string, block *bloomshipper.LazyBl
 	return archivePath, nil
 }
 
-// TODO common with block_downloader, extract under bloomutils
 func extractArchive(archivePath string, workingDirectoryPath string) error {
 	file, err := os.Open(archivePath)
 	if err != nil {
